requirements.txt
asgiref==3.5.2
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
django-rest-extensions==0.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
drf-spectacular-sidecar==2022.12.1
idna==3.4
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
lxml==4.9.2
MarkupSafe==2.1.2
MechanicalSoup==1.2.0
mysqlclient==2.1.1
openapi-codec==1.3.2
persiantools==3.0.1
psycopg2==2.9.6
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2022.6
PyYAML==6.0
requests==2.28.2
rest-framework-cache==0.1
simplejson==3.18.3
simplejwt==2.0.1
soupsieve==2.4
sqlparse==0.4.3
typing==3.7.4.3
typing_extensions==4.5.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.14
